Generalize XML Parser to allow for differnet language engines
